refactor(viewer): rename measurement to annotation and modify location
