+++ libs/insight-viewer/src/Annotation/hooks/useCircleAnnotation/utils/getCircleConnectingLine.spec.ts
