+++ libs/insight-viewer/src/Annotation/hooks/useCircleAnnotation/utils/getCircleTextPosition.spec.ts
