+++ libs/insight-viewer/src/Annotation/hooks/useCircleAnnotation/utils/getCircleTextPosition.ts
