+++ libs/insight-viewer/src/Annotation/hooks/useRulerMeasurement/index.ts
@@ -1,17 +1,17 @@
 import useTextBox from '../useTextBox'
-import { useOverlayContext } from '../../contexts'
+import { useOverlayContext } from '../../../contexts'
 
-import { modifyConnectingLine } from '../../utils/common/modifyConnectingLine'
-import { stringifyPoints } from '../../utils/common/stringifyPoints'
+import { modifyConnectingLine } from '../../../utils/common/modifyConnectingLine'
+import { stringifyPoints } from '../../../utils/common/stringifyPoints'
 import { getRulerTextPosition } from './utils/getRulerTextPosition'
 import { getRulerConnectingLine } from './utils/getRulerConnectingLine'
 
-import { HALF_OF_RULER_TEXT_BOX } from '../../const'
-import type { Point, RulerMeasurement } from '../../types'
+import { HALF_OF_RULER_TEXT_BOX } from '../../../const'
+import type { Point, RulerAnnotation } from '../../types'
 
 const formatRulerValue = (measuredValue: number, unit: string) => `${measuredValue.toFixed(1)}${unit}`
 
-const useRulerMeasurement = ({ startAndEndPoint, measuredValue, unit, textPoint }: RulerMeasurement) => {
+const useRulerAnnotation = ({ startAndEndPoint, measuredValue, unit, textPoint }: RulerAnnotation) => {
   const { pixelToCanvas } = useOverlayContext()
   const [textBox, ref] = useTextBox()
 
@@ -34,4 +34,4 @@ const useRulerMeasurement = ({ startAndEndPoint, measuredValue, unit, textPoint 
   return { rulerLine, ref, connectingLine, formattedValue, textBoxPoint, visibility }
 }
 
-export default useRulerMeasurement
+export default useRulerAnnotation
