+++ libs/insight-viewer/src/Annotation/hooks/useRulerMeasurement/utils/getRulerConnectingLine.spec.ts
