+++ libs/insight-viewer/src/Annotation/hooks/useRulerMeasurement/utils/getRulerConnectingLine.ts
