+++ libs/insight-viewer/src/Annotation/hooks/useRulerMeasurement/utils/getRulerTextPosition.spec.ts
